Refactor CasesCohortButton and SelectCohortsModal to integrate getStringOperands and getCaseIdsFromFilter functions
* Moved getStringOperands and getCaseIdsFromFilter logic directly into the respective components, enhancing clarity and maintainability.
* Removed the utils file as its functions are now directly implemented in the components, streamlining the codebase.
* Improved type safety and operand handling in both components for better consistency.
